add sb tests

+++ sdk/eventhub/azure-eventhub/azure/eventhub/_utils.py
@@ -49,7 +49,7 @@ TZ_UTC: timezone = timezone.utc
 CE_ZERO_SECONDS: int = -62_135_596_800
 
 
-def utc_from_timestamp(timestamp: int) -> datetime.datetime:
+def utc_from_timestamp(timestamp: float) -> datetime.datetime:
     """
     :param int timestamp: Timestamp in seconds to be converted to datetime.
     """
